torch backend use torch dataloader directly
